api/requirements.txt
geni-lib==0.9.9.4
Django==4.0.4
djangorestframework==3.13.1
simplejson==3.17.6
six==1.16.0
xmltodict==0.13.0
paramiko==2.11.0
scp==0.14.4